Fix Write to always return a value <= len(p)
According to the docs, Write must always return 0 <= n <= len(p), and bytes.Buffer panics if this is violated.
On the other hand it's now misreporting how many bytes it wrote which is also wrong, but often callers don't need to check n, and in my case it's easier to handle than a panic.

Possibly the most correct solution would be to panic if a non-even length was passed, but that would of course break existing users.